Added ATs for Number of Entries page and updated BaseStepDef
