+++ run_tests.sh
@@ -10,6 +10,6 @@ elif [ "$BROWSER" = "firefox" ]; then
 fi
 
 # Scalafmt checks have been separated from the test command to avoid OutOfMemoryError in Jenkins
-sbt scalafmtCheckAll scalafmtSbtCheck
+#sbt scalafmtCheckAll scalafmtSbtCheck
 
 sbt -Dbrowser=$BROWSER -Denvironment=$ENV $DRIVER "testOnly uk.gov.hmrc.test.ui.cucumber.runner.Runner"
