+++ src/test/scala/uk/gov/hmrc/test/ui/pages/BasePage.scala
@@ -16,8 +16,21 @@
 
 package uk.gov.hmrc.test.ui.pages
 
+import org.openqa.selenium.{By, WebElement}
 import org.scalatest.matchers.should.Matchers
+import uk.gov.hmrc.test.ui.driver.BrowserDriver
 
-trait BasePage extends Matchers {
-  val url: String
+trait BasePage extends Matchers with BrowserDriver {
+
+  // ID Wrappers
+  def findById(id: String) = findBy(By.id(id))
+  def clickById(id: String) = findById(id).click()
+
+  // Find By Syntax Sugar
+  def findBy(by: By) = driver.findElement(by)
+  def findElementsBy(by: By) = driver.findElements(by)
+  def findElementById(id: String): WebElement = driver.findElement(By.id(id))
+  def findElementByName(name: String): WebElement = driver.findElement(By.name(name))
+
+
 }
